\name{tth-package}
\alias{tth-package}

\title{R Interface to the tth/ttm TeX to HTML Converter}

\description{ tth/ttm are command line utilities written by Hutchinson
  (2012) for converting (La)TeX to HTML or HTML+MathML, respectively.
}

\details{ The R package \pkg{tth} ships the C sources for convenient
  compilation and installation on all platforms. It also provides wrappers in
  R to process R character vectors with the command line tools directly
  from the R prompt. A detailed manual for tth/ttm is available online at
  \url{http://hutchinson.belmont.ma.us/tth/}.
}

\references{
  Hutchinson IH (2012). \emph{TtH: The TeX to HTML Translator}, Version 4.03.
  \url{http://hutchinson.belmont.ma.us/tth/}
}

\keyword{utilities}
