\name{tth-package}
\alias{tth-package}

\title{C Sources of the tth/ttm TeX to HTML Converter}

\description{
  tth/ttm are C functions written by Hutchinson (2012) for converting (La)TeX to
  HTML or HTML+MathML, respectively.
}

\details{
  The R package \pkg{tth} just ships the sources for convenient compilation
  on all platforms. R interfaces to the C functions are provided in package
  \pkg{htmltools}. A detailed manual is available online
  at \url{http://hutchinson.belmont.ma.us/tth/manual.cgi} (which actually produces
  an HTML version of the manual in real time using tth).
}

\references{
  Hutchinson IH (2012). \emph{TtH: The TeX to HTML Translator}, Version 4.03.
  \url{http://hutchinson.belmont.ma.us/tth/}
}

\keyword{utilities}
